refactor: rename modules name for convention
